LRCI-3783 Move getModifiedDownstreamBuilds into ParentBuild
