+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/BaseBuild.java
@@ -873,30 +873,6 @@ public abstract class BaseBuild implements Build {
 	}
 
 	@Override
-	public List<Build> getModifiedDownstreamBuilds() {
-		return getModifiedDownstreamBuildsByStatus(null);
-	}
-
-	@Override
-	public List<Build> getModifiedDownstreamBuildsByStatus(String status) {
-		List<Build> modifiedDownstreamBuilds = new ArrayList<>();
-
-		for (Build downstreamBuild : downstreamBuilds) {
-			if (downstreamBuild.isBuildModified() ||
-				downstreamBuild.hasModifiedDownstreamBuilds()) {
-
-				modifiedDownstreamBuilds.add(downstreamBuild);
-			}
-		}
-
-		if (status != null) {
-			modifiedDownstreamBuilds.retainAll(getDownstreamBuilds(status));
-		}
-
-		return modifiedDownstreamBuilds;
-	}
-
-	@Override
 	public Map<String, String> getParameters() {
 		return new HashMap<>(_parameters);
 	}
